+++ plugins/source/airtable/src/plugin.ts
@@ -64,9 +64,16 @@ export const newAirtablePlugin = () => {
   const newClient: NewClientFunction = async (logger, spec, { noConnection }) => {
     pluginClient.spec = parseSpec(spec);
     pluginClient.client = { id: () => 'airtable' };
-    pluginClient.allTables = noConnection
-      ? []
-      : await getTables(pluginClient.spec.apiKey, pluginClient.spec.endpointUrl, pluginClient.spec.concurrency);
+    if (noConnection) {
+      pluginClient.allTables = [];
+      return pluginClient;
+    }
+    pluginClient.allTables = await getTables(
+      logger,
+      pluginClient.spec.apiKey,
+      pluginClient.spec.endpointUrl,
+      pluginClient.spec.concurrency,
+    );
 
     return pluginClient;
   };