+++ plugins/source/airtable/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.1.0](https://github.com/cloudquery/cloudquery/compare/plugins-source-airtable-v1.0.1...plugins-source-airtable-v1.1.0) (2023-08-18)
+
+
+### Features
+
+* Improve logging of discovery phase ([#13179](https://github.com/cloudquery/cloudquery/issues/13179)) ([5c7776f](https://github.com/cloudquery/cloudquery/commit/5c7776f332bc6a257f216b22a1fceefbca46357b))
+
+## [1.0.1](https://github.com/cloudquery/cloudquery/compare/plugins-source-airtable-v1.0.0...plugins-source-airtable-v1.0.1) (2023-08-17)
+
+
+### Bug Fixes
+
+* Airtable `dateTime` type handling ([#13174](https://github.com/cloudquery/cloudquery/issues/13174)) ([275a526](https://github.com/cloudquery/cloudquery/commit/275a526ea19fb166df816a22f0818188365c4a65))
+
 ## 1.0.0 (2023-08-17)